adding utilities, and chaning selections

+++ python/ZCandSelector_cfi.py
@@ -3,7 +3,7 @@ import FWCore.ParameterSet.Config as cms
 defaultZCandSelectionParameters = cms.PSet(
     massMin = cms.double(60),
     massMax = cms.double(120),
-    ptMin = cms.double(150),
+    ptMin = cms.double(0),
     ptMax = cms.double(100000), 
     )
